Update your profile added

+++ src/main/java/com/example/kinoarena/controllers/CinemaController.java
@@ -17,32 +17,42 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.kinoarena.dto.AddCinemaDto;
+import com.example.kinoarena.exceptions.CinemaNotFoundException;
 import com.example.kinoarena.exceptions.KinoArenaException;
+import com.example.kinoarena.helper.UserValidation;
 import com.example.kinoarena.model.Cinema;
+import com.example.kinoarena.model.Location;
 
 @RestController
 public class CinemaController extends BaseController {
 
+	private static final String CINEMA_NOT_FOUND = "There is no such cinema!";
+
+	UserValidation validation = new UserValidation();
+
 	@Autowired
 	private CinemaRepository cinemaRepository;
+
+	@Autowired
+	private LocationRepository locationRepository;
 
 	// IVANA CODE
 	@GetMapping("/cinema/{id}")
 	public Cinema chooseCinema(@PathVariable("id") Long id, HttpSession session, HttpServletResponse response,
 			HttpServletRequest request) throws KinoArenaException {
 		validateLogin(session);
-			Optional<Cinema> cin = cinemaRepository.findById(id);
-			Cinema cinema = null;
-			if (cin.isPresent()) {
-				cinema = cin.get();
-			}
-			if (cinema != null) {
-				session.setAttribute("cinema_id", id);
-				return cinema;
-			} else {
-				response.setStatus(404);
-				throw new KinoArenaException("There is not a cinema with id = " + id);
-			}
+		Optional<Cinema> cin = cinemaRepository.findById(id);
+		Cinema cinema = null;
+		if (cin.isPresent()) {
+			cinema = cin.get();
+		}
+		if (cinema != null) {
+			session.setAttribute("cinema_id", id);
+			return cinema;
+		} else {
+			response.setStatus(404);
+			throw new KinoArenaException("There is not a cinema with id = " + id);
+		}
 	}
 
 	// IVANA CODE
@@ -58,29 +68,47 @@ public class CinemaController extends BaseController {
 		}
 	}
 
-	// NOT WORKING CHECH LOCATIONID
+	// !WORKING
+	/**
+	 * Method for adding new cinema to a new location
+	 */
 	@PostMapping("/addcinema")
 	public void addCinema(@RequestBody AddCinemaDto cinema, HttpServletRequest request)
 			throws KinoArenaException, SQLException {
 		validateLoginAdmin(request);
-		Cinema c = new Cinema();
-		c.setContact(cinema.getContact());
-		c.setLocationId(cinema.getLocatonId());
-		c.setName(cinema.getName());
-		cinemaRepository.save(c);
+
+		Location location = new Location();
+	
+			validation.validateCityOrAddress(cinema.getCity());
+			location.setCity(cinema.getCity());
+			validation.validateCityOrAddress(cinema.getAddress());
+			location.setAddress(cinema.getAddress());
+			locationRepository.save(location);
+
+			Cinema kino = new Cinema();
+			kino.setLocationId(location.getLocationId());
+			validation.validateGsm(cinema.getContact());
+			kino.setContact(cinema.getContact());
+			validation.validateString(cinema.getName());
+			kino.setName(cinema.getName());
+			cinemaRepository.save(kino);
 	}
 
-	// NOT WORKING - Cannot delete or update a parent row: a foreign key constraint
-	// fails
+	// Working!
 	@DeleteMapping("/deletecinema/{id}")
 	public void deleteCinema(@PathVariable Long id, HttpServletRequest request) throws KinoArenaException {
 		validateLoginAdmin(request);
-		cinemaRepository.deleteById(id);
+		if (cinemaRepository.existsById(id)) {
+			cinemaRepository.deleteById(id);
+		} else {
+			throw new CinemaNotFoundException(CINEMA_NOT_FOUND);
+		}
 	}
 
 	// WORKING
 	@GetMapping("/cinemas")
-	public List<Cinema> getAll(HttpSession session, HttpServletRequest request) throws KinoArenaException, SQLException {
+	public List<Cinema> getAll(HttpSession session, HttpServletRequest request)
+			throws KinoArenaException, SQLException {
 		validateLogin(session);
 		return cinemaRepository.findAll();
 	}
